Define an app template and use it for the editor and chat apps
