+++ src/elva/commands/chat/style.tcss
@@ -11,7 +11,6 @@
 #tabview {
   height: 30%;
   min-height: 10;
-  dock: bottom;
 }
 
 .message {
@@ -36,37 +35,3 @@ Rule {
   margin: 0;
 }
 
-Dashboard {
-  layout: grid;
-  grid-size: 2;
-  grid-columns: 2fr 1fr;
-
-  ConfigView, AwarenessView {
-    padding: 0 1 0 1;
-    border: solid $primary;
-  }
-}
-
-InputScreen {
-  layout: grid;
-  align: center middle;
-
-  Input {
-    border: solid $border;
-    width: 80%;
-
-    .input--cursor {
-      color: $foreground;
-    }
-  }
-}
-
-ErrorScreen {
-  align: center middle;
-  border: solid $error;
-  
-  * {
-    content-align: center middle;
-    padding: 1 0 1 0;
-  }
-}
