+++ src/elva/commands/editor/style.tcss
@@ -6,52 +6,3 @@ YTextArea {
   background: transparent;
   height: 1fr;
 }
-
-Footer {
-  dock: bottom;
-}
-
-Dashboard {
-  layout: grid;
-  grid-size: 2;
-  grid-columns: 2fr 1fr;
-
-  ConfigView, AwarenessView {
-    padding: 0 1 0 1;
-    border: solid $primary;
-  }
-}
-
-RoomBrowserScreen {
-  layout: vertical;
-
-  DataTable {
-    height: 1fr;
-    padding: 0 1;
-    border: solid $primary;
-  }
-}
-
-InputScreen {
-  layout: grid;
-  align: center middle;
-
-  Input {
-    border: solid $border;
-    width: 80%;
-
-    .input--cursor {
-      color: $foreground;
-    }
-  }
-}
-
-ErrorScreen {
-  align: center middle;
-  border: solid $error;
-  
-  * {
-    content-align: center middle;
-    padding: 1 0 1 0;
-  }
-}
